In progress

+++ Mac/clearPlayerDebugMode.command
@@ -9,24 +9,27 @@ export projectHomeDir=`dirname "$scriptDir"`
 
 cd "$projectHomeDir"
 
-export CEP_VERSION=`head -n 1 "$projectHomeDir/CEPVersion.txt"`
+export CEP_VERSION=`head -n 1 "$projectHomeDir/BuildSettings/CEPVersion.txt"`
 
-if [ $CEP_VERSION = "4.x" ]; then
+if [ "$CEP_VERSION" == "4.x" ]; then
   export PLIST_FILE=com.adobe.CSXS.4.plist
-elif [ $CEP_VERSION = "5.x" ]; then
+elif [ "$CEP_VERSION" == "5.x" ]; then
   export PLIST_FILE=com.adobe.CSXS.5.plist
-elif [ $CEP_VERSION = "6.x" ]; then
+elif [ "$CEP_VERSION" == "6.x" ]; then
   export PLIST_FILE=com.adobe.CSXS.5.plist
-elif [ $CEP_VERSION = "7.x" ]; then
+elif [ "$CEP_VERSION" == "7.x" ]; then
   export PLIST_FILE=com.adobe.CSXS.5.plist
-elif [ $CEP_VERSION = "8.x" ]; then
+elif [ "$CEP_VERSION" == "8.x" ]; then
   export PLIST_FILE=com.adobe.CSXS.8.plist
 fi
 
-defaults write $PLIST_FILE PlayerDebugMode 0
-defaults write $PLIST_FILE LogLevel 1
+if [ "$PLIST_FILE" != "" ]; then
 
-#
-# Force Mac OS to re-read the PLIST file
-#
-killall -u `whoami` cfprefsd
+  defaults write $PLIST_FILE PlayerDebugMode 0
+  defaults write $PLIST_FILE LogLevel 1
+
+  #
+  # Force Mac OS to re-read the PLIST file
+  #
+  killall -u `whoami` cfprefsd
+fi
